+++ packages/waggle/jestSetup.js
@@ -1,0 +1,5 @@
+const { Crypto } = require('@peculiar/webcrypto')
+
+console.log('Custom jest setup')
+const crypto = new Crypto();
+global.crypto = crypto;