+++ packages/waggle/src/common/testUtils.ts
@@ -1,4 +1,5 @@
-import fp from 'find-free-port'
+import getPort from 'get-port'
+import Libp2p from 'libp2p'
 import { HttpsProxyAgent } from 'https-proxy-agent'
 import { Response } from 'node-fetch'
 import path from 'path'
@@ -7,7 +8,6 @@ import tmp from 'tmp'
 import { ConnectionsManagerOptions, DataFromPems } from '../common/types'
 import { Config } from '../constants'
 import { ConnectionsManager } from '../libp2p/connectionsManager'
-import { Libp2pType } from '../libp2p/customLibp2p'
 import { createCertificatesTestHelper } from '../libp2p/tests/client-server'
 import WebsocketsOverTor from '../libp2p/websocketOverTor'
 import logger from '../logger'
@@ -65,8 +65,8 @@ export const createMinConnectionManager = (options: ConnectionsManagerOptions): 
   })
 }
 
-export const createLibp2p = async (peerId: PeerId): Promise<Libp2pType> => {
-  const [port] = await fp(1111)
+export const createLibp2p = async (peerId: PeerId): Promise<Libp2p> => {
+  const port = await getPort()
   const virtPort = 443
   const pems = await createCertificatesTestHelper('address1.onion', 'address2.onion')