+++ packages/waggle/src/storage/storageSnapshot.ts
@@ -1,13 +1,14 @@
 import fs from 'fs'
 import Log from 'ipfs-log'
 // import { CID } from 'multiformats/cid'
+import Libp2p from 'libp2p'
 import OrbitDB from 'orbit-db'
 import EventStore from 'orbit-db-eventstore'
 import PeerId from 'peer-id'
 import { StorageOptions } from '../common/types'
+import { createPaths } from '../common/utils'
 import logger from '../logger'
 import { Storage } from '../storage'
-import { createPaths } from '../common/utils'
 
 const log = logger('dbSnap')
 
@@ -51,7 +52,7 @@ export class StorageTestSnapshot extends Storage {
     this.name = (Math.random() + 1).toString(36).substring(7)
   }
 
-  public async init(libp2p: any, peerID: PeerId): Promise<void> {
+  public async init(libp2p: Libp2p, peerID: PeerId): Promise<void> {
     log(`${this.name}; StorageTest: Entered init`)
     if (this.options?.createPaths) {
       createPaths([this.ipfsRepoPath, this.orbitDbDir])
@@ -59,6 +60,20 @@ export class StorageTestSnapshot extends Storage {
     this.ipfs = await this.initIPFS(libp2p, peerID)
 
     this.orbitdb = await OrbitDB.createInstance(this.ipfs, { directory: this.orbitDbDir })
+
+    await this.createDbForSnapshotInfo()
+    await this.createDbForMessages()
+    log(`Initialized '${this.name}'`)
+  }
+
+  public setName(name) {
+    this.name = name
+  }
+
+  private async createDbForSnapshotInfo() {
+    if (!this.useSnapshot) {
+      return
+    }
     this.snapshotInfoDb = await this.orbitdb.log<SnapshotInfo>('092183012', {
       accessController: {
         write: ['*']
@@ -67,8 +82,6 @@ export class StorageTestSnapshot extends Storage {
 
     // eslint-disable-next-line
     this.snapshotInfoDb.events.on('replicated', async () => {
-      if (!this.useSnapshot) return
-
       // Retrieve snapshot that someone else saved to db
       if (!this.options.createSnapshot || process.env.CREATE_SNAPSHOT !== 'true') {
         log('Replicated snapshotInfoDb')
@@ -82,12 +95,6 @@ export class StorageTestSnapshot extends Storage {
     // log(`${this.name}; replication in progress:`, address, hash, entry, progress, total)
     // log('>>', entry.payload.value.snapshot)
     // })
-    await this.createDbForMessages()
-    log(`Initialized '${this.name}'`)
-  }
-
-  public setName(name) {
-    this.name = name
   }
 
   private async createDbForMessages() {